tiny aesthetic change + to-do list
forgot to make the tags small last time

also removed right padding from .reg__details (allows a full author name
to fit on one line when byline/date are in the left column). I tried
expanding .reg__details to 3 columns instead of 2, but it looks pretty
big and empty so I changed it back and did this as compromise. not sure
how I feel about it aesthetically though, might change back tomorrow and
try something else.

to-do list:
- put the actual elements we want (top articles, creative commons) in
the sidebar
- test styles for sidebar
- media query so only creative commons part of sidebar stacks on mobile
- fix styling for .reg__details at md breakpoint
- style different heading sizes, pullquotes, embedded images etc.
- should probably style footnotes at some point