Installed typi
Installed typi and added the basic sass maps it runs on. To do next:
calculate the actual modular scale for all the type on the site and add
that to the typi maps.

A potential problem - typi doesn't play entirely nicely with the way
neat defines breakpoints. I've worked around it for the time being by
defining the neat breakpoints based on the typi breakpoints map, which
at least keeps everything consistent, though it does mean there are now
two places in the variables file where breakpoints are added/deleted.