styled share
changed markup in share fragment so that only the circle is a link.
added a hover effect to the circle.

at the moment the SVGs aren't perfectly centred when the breakpoint size
changes. I'll fix this l8r.

getting the text to centre vertically is also a little hacky, will also
return to later