fixed width of related
it's now the same width as the main column

the 1px border on the related items is less for style and more for me to
see the grid as I work on it. running into problems getting the boxes in
this row to be the same height. there are some dodgy clearfix solutions
but I think we need to look into flexbox