hacky fix for most popular widget + single article widget
for some reason multi-line list items on the most popular list were
doing all kinds of weird shit i.e. indenting strangely, not sitting
inline with the number etc. I've tried all kinds of fixes and I'm
stumped for the moment...I imagine it's a pretty simple error and it
will come to me at some point soon. For the time being this is a hacky
fix that works, spacing/style not final though.

There's also some mucking around on the single-article widget styling or
something