header & footer style mockup
this is a big commit sorry hope it doesn't #breaktheinternet (in a bad
way not a kardashian way)

some notes
- this is very unfinished, just thought I'd commit the general gist
- I have only styled the large breakpoint so far. I know this is not
mobile-first this was purely for the purpose of a mockup.
- font size, font spacing, font color, contrast (!!) need to be worked
on heaps
- my ~vision~ for the search box is just the search SVG and then u click
and it expands a minimalist search box similar to the above (text field
is just a line w/ the search symbol next to it)
- spacing and footer layout is not done
- I have coded this messily by commenting out lots of stuff instead of
deleting, and have just written these styles in the large breakpoint. If
we like the general direction I'll clean it up.
- there are light and dark versions. the one I'm committing is the dark
one, but if you go into the bottom of the variables file and comment out
the variables as directed you can see the light version. the colours
aren't right but it's a general gist