mucking around with logo size
plus misc small other stuff. added a proper size for the logo to smaller
breakpoints and enlarged it slightly on the large size in the hope of
appeasing MW. not happy about it tho, considering trying it all in a bar
w/ logo to left of nav and seeing if that looks less weirdly small to
her?