better archive page - large breakpoints
have overhauled the archive page bc it looked like trash. I think it is
approaching something better now.

I've only sorted the largest breakpoint, the smaller ones are a bit shit
rn.

Potential problems: the category and primary tag seeming so large
relative to everything, including being in a box. The rationale for the
box was that it forms a nice grid w/ the top of the image in a way the
tag alone does not. Have tried to make it a bit unobtrusive.

I also envisage the author names only having the coloured underline
rather than coloured text, but idk how to do that yet