gradients + blockquote styling
made a thing where you can @include gradient-background on something and
it will automatically generate a CSS gradient from the $brand-primary
and $brand-secondary colours you define in variables. I think it's
pretty ~fun~/a bit bloomberg, but it's easy to disable if you hate

also styled the blockquote. it's boring but fine I think, as blockquotes
should be. change shit if you want