Correct display of archive article summaries
Pages returned in search results were loading the whole pages markup as
the summary. This was uncool, so I’ve simplified such that the same
article summary (in content.php) is used to display search result
entries for both posts and pages.

Note that pages generally don’t have standfirsts or featured images so
their display is a little sparse. Not a big deal, but note have been
added to the to do spreadsheet so we consider it later